+++ src/main/java/ddd/caffeine/ratrip/common/exception/ExceptionResponse.java
@@ -1,4 +1,4 @@
-package ddd.caffeine.ratrip.exception;
+package ddd.caffeine.ratrip.common.exception;
 
 import org.springframework.http.HttpStatus;
