+++ src/main/java/ddd/caffeine/ratrip/common/exception/domain/KakaoFeignException.java
@@ -1,7 +1,7 @@
-package ddd.caffeine.ratrip.exception.domain;
+package ddd.caffeine.ratrip.common.exception.domain;
 
-import ddd.caffeine.ratrip.exception.BaseException;
-import ddd.caffeine.ratrip.exception.ExceptionInformation;
+import ddd.caffeine.ratrip.common.exception.BaseException;
+import ddd.caffeine.ratrip.common.exception.ExceptionInformation;
 import lombok.Getter;
 
 @Getter
