--- src/main/java/ddd/caffeine/ratrip/controller/dto/RecommendRequestDto.java
@@ -1,4 +1,0 @@
-package ddd.caffeine.ratrip.controller.dto;
-
-public class RecommendRequestDto {
-}
