--- src/main/java/ddd/caffeine/ratrip/controller/dto/RecommendResponseDto.java
@@ -1,4 +1,0 @@
-package ddd.caffeine.ratrip.controller.dto;
-
-public class RecommendResponseDto {
-}
