+++ src/main/java/ddd/caffeine/ratrip/module/recommend/presentation/RecommendPlaceController.java
@@ -1,4 +1,4 @@
-package ddd.caffeine.ratrip.controller;
+package ddd.caffeine.ratrip.module.recommend.presentation;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -6,8 +6,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import ddd.caffeine.ratrip.service.RecommendPlaceUseCase;
-import ddd.caffeine.ratrip.service.feign.model.KakaoFeignResponseDto;
+import ddd.caffeine.ratrip.module.recommend.presentation.dto.RecommendResponseDto;
+import ddd.caffeine.ratrip.module.recommend.service.RecommendPlaceUseCase;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 
@@ -22,10 +22,11 @@ public class RecommendPlaceController {
 	private final RecommendPlaceUseCase recommendPlaceUseCase;
 
 	@GetMapping(value = "place")
-	public ResponseEntity<KakaoFeignResponseDto> callRecommendPlaceApi(@RequestParam String region,
+	public ResponseEntity<RecommendResponseDto> callRecommendPlaceApi(@RequestParam String region,
 		@RequestParam String keyword, @RequestParam(required = false, defaultValue = "1") int page
 	) {
-		KakaoFeignResponseDto response = recommendPlaceUseCase.recommendPlace(region, keyword, page);
+		RecommendResponseDto response = new RecommendResponseDto(
+			recommendPlaceUseCase.recommendPlace(region, keyword, page));
 		return ResponseEntity.ok(response);
 	}
 }
