+++ src/main/java/ddd/caffeine/ratrip/module/recommend/presentation/dto/RecommendResponseDto.java
@@ -1,0 +1,21 @@
+package ddd.caffeine.ratrip.module.recommend.presentation.dto;
+
+import java.util.List;
+
+import ddd.caffeine.ratrip.module.recommend.domain.KakaoFeignData;
+import ddd.caffeine.ratrip.module.recommend.domain.KakaoFeignMetaData;
+import ddd.caffeine.ratrip.module.recommend.domain.KakaoFeignModel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class RecommendResponseDto {
+	private List<KakaoFeignData> documents;
+	private KakaoFeignMetaData meta;
+
+	public RecommendResponseDto(KakaoFeignModel kakaoFeignModel) {
+		this.documents = kakaoFeignModel.getDocuments();
+		this.meta = kakaoFeignModel.getMeta();
+	}
+}
