+++ src/main/java/ddd/caffeine/ratrip/module/recommend/service/RecommendPlaceUseCase.java
@@ -1,0 +1,7 @@
+package ddd.caffeine.ratrip.module.recommend.service;
+
+import ddd.caffeine.ratrip.module.recommend.domain.KakaoFeignModel;
+
+public interface RecommendPlaceUseCase {
+	KakaoFeignModel recommendPlace(String region, String keyword, int page);
+}
